refactor: consolidate key types into libapp
